+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Added
 Changed
  - You can now give default values or an array when initializing a
    `tcod.console.Console` instance.
+Deprecated
+ - Most libtcodpy console functions have been replaced by the tcod.console
+   module.
 
 8.4.3 - 2019-02-06
 ------------------